requirements.txt
backports.shutil-get-terminal-size (1.0.0)
decorator (4.1.2)
enum34 (1.1.6)
functools32 (3.2.3.post2)
ipdb (0.10.3)
ipython (5.5.0)
ipython-genutils (0.2.0)
pathlib2 (2.3.0)
pexpect (4.2.1)
pickleshare (0.7.4)
pip (9.0.1)
prompt-toolkit (1.0.15)
ptyprocess (0.5.2)
Pygments (2.2.0)
PyYAML (3.12)
scandir (1.6)
setuptools (36.6.0)
simplegeneric (0.8.1)
six (1.11.0)
traitlets (4.3.2)
wcwidth (0.1.7)
wheel (0.30.0)
